Created program for robotic arm using formulas
This program operates a robotic arm of servo motors based on input from four potentiometers.

Functions of each potentiometer:
        0 - X Position of Endpoint
        1 - Y Position of Endpoint
        2 - Claw Open-Close
        3 - Arm Rotation at base

Note:  When using this program, it is strongly recommended that you use a 7V 3A power source rather than a simple USB port.  Connect this power source to both the servo motors and the VIN port on the Arduino.  If you power the servo motors from the 5V ports on the Arduino board, they may not receive enough power to hold the robotic arm up.

This program is verified by the Arduino compiler, but it has not been tested yet.