Eliminated error
The error was in the line:
    return;
in both the alpha and beta functions.  The error was that this line had nothing to return, as both the alpha and beta functions are supposed to return a double type value.  This line was run when the the point (c, d) was not not possible to reach (Solution out of range).  For example, if the arm can stretch 10 units away from the base and you tried to use these functions to calculate the angles in the arm needed to reach 100 units away from the base, the function would return nothing.  However, the C++ compiler complains about this line of code, so it was eliminated, as well as the else statement that enclosed it.  Now, the functions compile without error and the functions still return nothing when the solution is out of range because there is no line of code to return anything at all in that case.