Organized Code and Fixed Error
Eliminated ".0" from ends of numbers.  Although they clarified that the value was a double, they are not necessary in this case.
Also, there was an error in the alpha function.  Although it worked correctly in most cases, it would return a strange value in others.  The reason for this was this:

Both the alpha and beta functions have the following two lines of code before returning an angle value:
    while (round(angle) >= 360) { angle -= 360; }
    while (round(angle) < 0) { angle += 360; }
These lines ensure that the angle being returned is between 0 and 360 degrees.  For example, we wouldn't want the function to return -700 when it should really return 20, but the calculations can make things like this possible, especially when c < 0.  The error was that, when these functions were run, the angle measure was still in radians, not degrees, so adding or subtracting 360 degrees was really more like adding or subtracting about 57.3 full rotations, and who knows where that might end up!  The angle measure was then converted to degrees in the return statement at the end of the function.  Therefore, if the radian value was already between 0 and 360, then there was no problem.  When it was outside the range, like -2, then the radian value would get 360 added to it.  The new radian value would then be converted to degrees, but it would no longer be coterminal with the original calculation.

The solution was simple:  Convert radians to degrees right as the angle variable is getting its radian value rather than at the end, when the value is returned.

✅  This code is verified by the C++ compiler.
❌  This code has not been tested on the robotic arm, yet.